pyproject.toml
[tool.poetry]
name = "mekara"
version = "0.1.0"
description = "Claude Agent SDK CLI"
authors = ["Meksys"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "mekara", from = "src" },
]
exclude = ["docs/"]
homepage = "https://github.com/meksys-dev/mekara"
keywords = ["claude", "agent", "sdk"]

[tool.poetry.dependencies]
python = "^3.12"
mcp = { version = ">=1.1.0", extras = ["cli"] }
pyyaml = ">=6.0"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.5.6"
pyright = ">=1.1.379"
pytest = ">=8.0"
pytest-asyncio = "^1.3.0"
pre-commit = ">=3.5.0"
syrupy = "^4.0.0"
vulture = "^2.14"

[tool.poetry.scripts]
mekara = "mekara.cli:main"

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pyright]
pythonVersion = "3.12"
include = [".mekara", "src", "tests"]
exclude = ["src/mekara/bundled/scripts"]  # Auto-generated scripts (both compiled and nl)

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "requires_llm: tests that make real LLM API calls"
]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
